Complete API

+++ MarsRoverPhotos.Services/Service/FileService.cs
@@ -25,20 +25,24 @@ namespace MarsRoverPhotos.Services.Service
         public async Task SaveImageToPath(Photo photo)
         {
             var uri = new Uri(photo.Img_Src);
+
+            var datePath = $"{path}\\{photo.Earth_Date:yyyy-MM-dd}\\";
+            var completePath = Path.Combine(datePath, $"{photo.Camera.Name}-{photo.Id}.JPG");
+
             var requestMessage = new HttpRequestMessage(HttpMethod.Get, photo.Img_Src);
             var response = await _httpClient.SendAsync(requestMessage);
-
             if (!response.IsSuccessStatusCode)
             {
                 throw new FileNotFoundException("File path not found!");
             }
+
+            if (File.Exists(completePath))
+            {
+                return;
+            }
             var imageBytes = await _httpClient.GetByteArrayAsync(uri);
 
-            // Create file path and ensure directory exists
-            var datePath = $"{path}\\{photo.Earth_Date:yyyy-MM-dd}\\";
-            var completePath = Path.Combine(datePath, $"{photo.Camera.Name}-{photo.Id}.JPG");
             Directory.CreateDirectory(datePath);
-
             await File.WriteAllBytesAsync(completePath, imageBytes);
         }
 
@@ -60,11 +64,47 @@ namespace MarsRoverPhotos.Services.Service
             return dateList;
         }
 
+        public async Task<byte[]> GetPhotoStream(DateTime date, string fileName)
+        {
+            var filePath = $"{path}\\{date:yyyy-MM-dd}\\{fileName}";
+            if (!File.Exists(filePath))
+            {
+                return null;
+            }
+
+            return await File.ReadAllBytesAsync(filePath);
+        }
+
         public IList<string> AvailableDates()
         {
-            var dates = Directory.GetDirectories(path).ToList();
+            if (!Directory.Exists(path))
+            {
+                return null;
+            }
+            List<string> dates = Directory.GetDirectories(path)
+                .Select(x => x.Split('\\').LastOrDefault())
+                .OrderBy(x => x).ToList();
 
             return dates;
+        }
+
+        public Domain.Entities.Dto.PhotoList AvailablePhotos(string datePath)
+        {
+            var fullPath = $"{ path}\\{ datePath}";
+            if (!Directory.Exists(fullPath))
+            {
+                return null;
+            }
+
+            var photoList = new Domain.Entities.Dto.PhotoList
+            {
+                Photos = Directory.GetFiles(fullPath)
+                    .Select(x => x.Split('\\').LastOrDefault())
+                    .OrderBy(x => x).ToList(),
+                DateString = datePath
+            };
+
+            return photoList;
         }
     }
 }
